requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
opencv-contrib-python==4.8.1.78
pandas==2.2.3
numpy==1.26.4
Pillow==10.1.0